+++ server/graphql/programs/src/lib.rs
@@ -5,11 +5,7 @@ use graphql_core::pagination::PaginationInput;
 use graphql_core::standard_graphql_error::validate_auth;
 use graphql_core::standard_graphql_error::StandardGraphqlError;
 use graphql_core::ContextExt;
-use graphql_types::types::document::DocumentFilterInput;
 use graphql_types::types::document::DocumentNode;
-use graphql_types::types::document::DocumentSortInput;
-use graphql_types::types::document_registry::DocumentRegistryFilterInput;
-use graphql_types::types::document_registry::DocumentRegistrySortInput;
 use graphql_types::types::encounter::EncounterFilterInput;
 use graphql_types::types::encounter::EncounterSortInput;
 use graphql_types::types::patient::PatientNode;
