+++ server/repository/src/db_diesel/clinician_store_join_row.rs
@@ -21,6 +21,7 @@ table! {
 }
 
 joinable!(clinician_store_join -> clinician (clinician_id));
+allow_tables_to_appear_in_same_query!(clinician, clinician_store_join);
 
 #[derive(Clone, Queryable, Insertable, AsChangeset, Debug, PartialEq, Default)]
 #[table_name = "clinician_store_join"]
