+++ server/repository/src/db_diesel/diesel_schema.rs
@@ -1,11 +1,11 @@
 use crate::program;
 
 use super::{
-    barcode_row::barcode, clinician_row::clinician, clinician_store_join_row::clinician_store_join,
-    invoice_line::invoice_stats, invoice_line_row::invoice_line, invoice_row::invoice,
-    item_row::item, location_row::location, master_list_line_row::master_list_line,
-    master_list_name_join::master_list_name_join, master_list_row::master_list, name_row::name,
-    name_store_join::name_store_join, name_tag_join::name_tag_join, period::period,
+    barcode_row::barcode, clinician_row::clinician, invoice_line::invoice_stats,
+    invoice_line_row::invoice_line, invoice_row::invoice, item_row::item, location_row::location,
+    master_list_line_row::master_list_line, master_list_name_join::master_list_name_join,
+    master_list_row::master_list, name_row::name, name_store_join::name_store_join,
+    name_tag_join::name_tag_join, period::period,
     program_requisition::program_requisition_order_type_row::program_requisition_order_type,
     program_requisition::program_requisition_settings_row::program_requisition_settings,
     requisition_line_row::requisition_line, requisition_row::requisition,
@@ -39,5 +39,4 @@ allow_tables_to_appear_in_same_query!(
     name_tag_join,
     barcode,
     clinician,
-    clinician_store_join
 );
